+++ src/main/java/com/voidpulse/pulseevents/listener/PlayerSurvivalListener.java
@@ -1,6 +1,7 @@
 package com.voidpulse.pulseevents.listener;
 
 import com.voidpulse.pulseevents.manager.EventManager;
+import com.voidpulse.pulseevents.manager.ProgressionManager;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
@@ -9,9 +10,11 @@ import org.bukkit.event.player.PlayerQuitEvent;
 public class PlayerSurvivalListener implements Listener {
 
     private final EventManager eventManager;
+    private final ProgressionManager progressionManager;
 
-    public PlayerSurvivalListener(EventManager eventManager) {
+    public PlayerSurvivalListener(EventManager eventManager, ProgressionManager progressionManager) {
         this.eventManager = eventManager;
+        this.progressionManager = progressionManager;
     }
 
     @EventHandler
@@ -22,5 +25,9 @@ public class PlayerSurvivalListener implements Listener {
     @EventHandler
     public void onPlayerQuit(PlayerQuitEvent event) {
         eventManager.markPlayerFailedCurrentEvent(event.getPlayer());
+
+        if (progressionManager != null) {
+            progressionManager.saveAll();
+        }
     }
 }